[kernel] kbuild: AFTER_LINK
If the make variable AFTER_LINK is set, it is a command line to run
after each final link.  This includes vmlinux itself and vDSO images.

Bugzilla: N/A
Upstream-status: ??

Herton: note that this patch is necessary for building debuginfo packages. This
is the same patch used in latest Fedora 4.9 kernels, and a similar patch is
applied to RHEL 7 kernel. Without it, build with debuginfo will fail.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>