[redhat] config review: CONFIG_RDMA_RXE=m
Set CONFIG_RDMA_RXE=m, as pointed on feedback from Jarod Wilson, Doug
Ledford, Don Dutile. A request to mark the driver as "tech preview" was
also done, however that is out of scope of config review/changes, and
someone will need to deal/do whatever is necessary in that respect
later.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>