[redhat] config review: cleanup/fix stack protector settings
CONFIG_CC_STACKPROTECTOR_STRONG is the default/desired setting, and we
only need to set this choice, we already set that as a generic option.
Also aarch64 was selecting other choice, but we want to also enable
CONFIG_CC_STACKPROTECTOR_STRONG on it as well. Thus, cleanup everything
and keep only CONFIG_CC_STACKPROTECTOR_STRONG set (this in practice
enable it also on aarch64). ppc64le doesn't set/have
HAVE_CC_STACKPROTECTOR and this can't be enabled/selected on it.
Reviewed also by Dave Anderson and Linda Wang.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>