[redhat] config review: do not disable CEPH_FS on aarch64
This was still an open item, what action to do related to Ceph FS on Pegas.
Turns out Ceph team wants it enabled so far, and this was already enabled from
the config from RHEL 7, however inherited arm64 config from rhelsa had it
disabled, so what we need to do now is just avoiding disabling this on aarch64
and this way we keep ceph with same settings enabled on all arches.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>