[redhat] allow building binaries with different name from src.rpm
We want to produce binary packages with a different name from the name
of the source package (kernel-*.rpm vs kernel-pegas-*.src.rpm). Introduce
support in the spec file to be able to do this, and with macros to easily switch
naming if/when needed. This also reverts the name of the binary packages, from
kernel-pegas* etc. to kernel only, names of the binary packages produced
now matches the same names as in RHEL 7 kernel packages.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>