[s390x] add zfcpdump user space tools from RHEL 7
Message-id: <20170130002317.25517-3-herton@redhat.com>
Patchwork-id: 163692
O-Subject: [Pegas 1.0 PATCH v2 2/3] [s390x] add zfcpdump user space tools from RHEL 7
Bugzilla: 1410579
RH-Acked-by: Lauro Ramos Venancio <lvenanci@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Josh Boyer <jwboyer@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1410579
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=12453980

The zfcpdump user space tool is necessary to enable again the s390 build on Pegas
(used for the s390 kdump build flavour). It is copied as is from the current
RHEL 7 tree.

The sources added by this patch are from an user space tool (zfcpdump), not
kernel code. It goes into the builtin initramfs included with the kernel image
for the s390 kdump package (see CONFIG_INITRAMFS_SOURCE kernel config option
definition and setting used by it), the same way it could be included by eg.
dracut if INITRAMFS_SOURCE was not used and an split initramfs was created. I
believe the reason we have it this way (including the source and building
manually so it is available for the builtin initramfs, see the rpm spec file
which builds it) is because s390utils on RHEL doesn't provide an static zfcpdump
binary, so it was chosen previously to be included and built in the kernel
package build process (I reserved a s390 system on beaker to check, and I can't
see any package providing zfcpdump).

If it was not the case we could make the package BuildRequires the proper
s390utils package and change the zfcpdump_initramfs.txt file to point to the
zfcpdump binary installed by the package.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>