[clocksource] arm_arch_timer: split dt-only rate handling
Message-id: <20170427082814.15855-8-juma@redhat.com>
Patchwork-id: 177244
O-Subject: [PATCH Pegas1.0 BZ1403765 07/16] clocksource: arm_arch_timer: split dt-only rate handling
Bugzilla: 1403765
RH-Acked-by: Al Stone <ahs3@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Mark Langsdorf <mlangsdo@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Fu Wei <fu.wei@linaro.org>

https://git.kernel.org/pub/scm/linux/kernel/git/tip/tip.git/log/?h=timers/core
commit 5d3dfa96c7a202f3a70ebeb6f8bb5005c1547250

For historical reasons, rate detection when probing via DT is somewhat
convoluted. We tried to package this up in arch_timer_detect_rate(), but
with the addition of ACPI worse, and gets in the way of stringent rate
checking when ACPI is used.

This patch makes arch_timer_detect_rate() specific to DT, ripping out
ACPI logic. In preparation for rework of the MMIO timer probing, the
reading of the relevant CNTFRQ register is factored out to callers. The
function is then renamed to arch_timer_of_configure_rate(), which better
represents its new place in the world.

Comments are added in the DT and ACPI probe paths to explain this.

Signed-off-by: Fu Wei <fu.wei@linaro.org>
[Mark: reword commit message]
Signed-off-by: Mark Rutland <mark.rutland@arm.com>

Signed-off-by: Jun ma <juma@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>