[clocksource] arm_arch_timer: refactor arch_timer_needs_probing
Message-id: <20170427082814.15855-9-juma@redhat.com>
Patchwork-id: 177242
O-Subject: [PATCH Pegas1.0 BZ1403765 08/16] clocksource: arm_arch_timer: refactor arch_timer_needs_probing
Bugzilla: 1403765
RH-Acked-by: Al Stone <ahs3@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Mark Langsdorf <mlangsdo@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Fu Wei <fu.wei@linaro.org>

https://git.kernel.org/pub/scm/linux/kernel/git/tip/tip.git/log/?h=timers/core
commit 13bf6992fe5496d56101b5d57fbdae2ee91184f2

When booting with DT, it's possible for timer nodes to be probed in any
order. Some common initialisation needs to occur after all nodes have
been probed, and arch_timer_common_init() has code to detect when this
has happened.

This logic is DT-specific, and it would be best to factor it out of the
common code that will be shared with ACPI.

This patch folds this into the existing arch_timer_needs_probing(),
which is renamed to arch_timer_needs_of_probing(), and no longer takes
any arguments. This is only called when using DT, and not when using
ACPI, which will have a deterministic probe order.

Signed-off-by: Fu Wei <fu.wei@linaro.org>
Reviewed-by: Hanjun Guo <hanjun.guo@linaro.org>
[Mark: reword commit message]
Signed-off-by: Mark Rutland <mark.rutland@arm.com>

Signed-off-by: Jun ma <juma@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>