[arm64] vdso: Remove ISB from gettimeofday
Message-id: <1494004772-24064-3-git-send-email-rrichter@redhat.com>
Patchwork-id: 177690
O-Subject: [Pegas 1.0 BZ 1445440 2/2] arm64:vdso: Remove ISB from gettimeofday.
Bugzilla: 1445440
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>

From: Andrew Pinski <apinski@cavium.com>

ISB is normally required before mrs CNTVCT if we want the
mrs to completed after the loads. In this case it is not.
As we are taking the difference and if that difference
was going to be negative, we just use the last counter value
instead.

Signed-off-by: Andrew Pinski <apinski@cavium.com>
Message-Id: <1492991221-5156-2-git-send-email-apinski@cavium.com>
Signed-off-by: Robert Richter <rrichter@cavium.com>
Signed-off-by: Robert Richter <rrichter@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>