[misc] cxl: Keep track of mm struct associated with a context
Message-id: <20170508203821.8477.89758.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 177941
O-Subject: [PATCH PEGAS1.0 BZ1320907 3/7] cxl: Keep track of mm struct associated with a context
Bugzilla: 1320907
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1320907

Description:
------------
The mm_struct corresponding to the current task is acquired each time
an interrupt is raised. So to simplify the code, we only get the
mm_struct when attaching an AFU context to the process.
The mm_count reference is increased to ensure that the mm_struct can't
be freed. The mm_struct will be released when the context is detached.
A reference on mm_users is not kept to avoid a circular dependency if
the process mmaps its cxl mmio and forget to unmap before exiting.
The field glpid (pid of the group leader associated with the pid), of
the structure cxl_context, is removed because it's no longer useful.

Signed-off-by: Christophe Lombard <clombard@linux.vnet.ibm.com>
Reviewed-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Acked-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 6dd2d23403396d8e6d153a6c9db56e1a1012bad8

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=6dd2d23403396d8e6d153a6c9db56e1a1012bad8
----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>