[netdrv] hns: support deferred probe when no mdio
Message-id: <20170509234015.11475-22-zhowang@redhat.com>
Patchwork-id: 178023
O-Subject: [PATCH V2 Pegas1.0 BZ1266302 21/22] net: hns: support deferred probe when no mdio
Bugzilla: 1266302
RH-Acked-by: Mark Langsdorf <mlangsdo@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: lipeng <lipeng321@huawei.com>

git git://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git
commit 804ffe5c6197fc2696681c597ce096d93eb24af5

In the hip06 and hip07 SoCs, phy connect to mdio bus.The mdio
module is probed with module_init, and, as such,
is not guaranteed to probe before the HNS driver. So we need
to support deferred probe.

We check for probe deferral in the mac init, so we not init DSAF
when there is no mdio, and free all resource, to later learn that
we need to defer the probe.

Signed-off-by: lipeng <lipeng321@huawei.com>
Reviewed-by: Yisen Zhuang <yisen.zhuang@huawei.com>
Reviewed-by: Matthias Brugger <mbrugger@suse.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Zhou Wang <zhowang@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>