[scsi] hisi_sas: remove hisi_sas_port_deformed()
Message-id: <20170509234647.12290-7-zhowang@redhat.com>
Patchwork-id: 178033
O-Subject: [PATCH V2 Pegas1.0 BZ1266303 06/32] scsi: hisi_sas: remove hisi_sas_port_deformed()
Bugzilla: 1266303
RH-Acked-by: Mark Langsdorf <mlangsdo@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: John Garry <john.garry@huawei.com>

git git://git.kernel.org/pub/scm/linux/kernel/git/mkp/scsi.git
commit 405314df566c6168715413db682adeba297fd684

Currently when a root PHY is deformed from a asd_sas_port we try to
release the slots in the LLDD, and fail.

Regardless, it is not right to release this early.

This patch removes the deformed function. As it was before, port
deformation is still done in hisi_sas_phy_down().

It would be nice to actually remove the hisi_sas_port_{de}formed() pair,
however we cannot as we need to know the asd_sas_port index libsas has
associated with an asd_sas_phy.

The hw does actually generate a port id for a PHY, but this seems to a
random number, so ignored for this purpose.

This patch also changes the code to link slots to the hisi_sas_device,
and not hisi_sas_port.

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Zhou Wang <zhowang@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>