[iommu] of: Prepare for deferred IOMMU configuration
Message-id: <20170510090120.15052-3-juma@redhat.com>
Patchwork-id: 178107
O-Subject: [PATCH v2 Pegas1.0 BZ1266462 02/13] iommu/of: Prepare for deferred IOMMU configuration
Bugzilla: 1266462
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Robin Murphy <robin.murphy@arm.com>

commit d7b0558230e444f29488fcee0b0b561015d16f8a

IOMMU configuration represents unchanging properties of the hardware,
and as such should only need happen once in a device's lifetime, but
the necessary interaction with the IOMMU device and driver complicates
exactly when that point should be.

Since the only reasonable tool available for handling the inter-device
dependency is probe deferral, we need to prepare of_iommu_configure()
to run later than it is currently called (i.e. at driver probe rather
than device creation), to handle being retried, and to tell whether a
not-yet present IOMMU should be waited for or skipped (by virtue of
having declared a built-in driver or not).

Tested-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
Signed-off-by: Jun ma <juma@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>