[acpi] iort: Add function to check SMMUs drivers presence
Message-id: <20170510090120.15052-5-juma@redhat.com>
Patchwork-id: 178108
O-Subject: [PATCH v2 Pegas1.0 BZ1266462 04/13] ACPI/IORT: Add function to check SMMUs drivers presence
Bugzilla: 1266462
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>

commit 1d9029d440e40b276c0691caed1de10c42d96bef

The IOMMU probe deferral implementation requires a mechanism to detect
if drivers for SMMU components are built-in in the kernel to detect
whether IOMMU configuration for a given device should be deferred (ie
SMMU drivers present but still not probed) or not (drivers not present).

Add a simple function to IORT to detect if SMMU drivers for SMMU
components managed by IORT are built-in in the kernel.

Tested-by: Hanjun Guo <hanjun.guo@linaro.org>
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Reviewed-by: Robin Murphy <robin.murphy@arm.com>
Cc: Sricharan R <sricharan@codeaurora.org>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
Signed-off-by: Jun ma <juma@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>