[block] sg_io: introduce unpriv_sgio queue flag
Message-id: <20170512152022.22616-3-pbonzini@redhat.com>
Patchwork-id: 178301
O-Subject: [Pegas 1.0 PATCH 2/2] sg_io: introduce unpriv_sgio queue flag
Bugzilla: 1394238
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This queue flag will let unprivileged users send any SG_IO command to the
device, without any filtering.  This makes it possible to run a program
where you want to access the full range of SCSI commands, while still
running as confined as possible.  With this patch, such a program will
not need the CAP_SYS_RAWIO capability anymore.  It will also *not* be
able to send SCSI commands to a partition, or to a device mapper target
that does not map the entirety of the underlying block device.

The RHEL7 BZ had another patch, which has been resubmitted independently
and is now upstream (commit 25cdb64, "block: allow WRITE_SAME commands
with the SG_IO ioctl", 2016-12-19) after having been rejected when I
submitted it in 2012.  Because of this, I am going to try again getting
the patch upstream for 4.12.  However, it will hopefully have a more
powerful (but different) userspace API.  So for backwards compatibility
reasons Pegas is stuck with this non-upstream patch.

Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Ewan Milne <emilne@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>