[mm] migrate: new migrate mode MIGRATE_SYNC_NO_COPY
Message-id: <1495744346-12780-12-git-send-email-jglisse@redhat.com>
Patchwork-id: 178982
O-Subject: [Pegas 1.0 11/15] mm/migrate: new migrate mode MIGRATE_SYNC_NO_COPY
Bugzilla: 1433282 1422592
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Introduce a new migration mode that allow to offload the copy to
a device DMA engine. This changes the workflow of migration and
not all address_space migratepage callback can support this. So
it needs to be tested in those cases.

This is intended to be use by migrate_vma() which itself is use
for thing like HMM (see include/linux/hmm.h).

Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>