[arm] kvm: arm/arm64: Rework gpa callback handlers
Message-id: <1495824903-9192-24-git-send-email-wei@redhat.com>
Patchwork-id: 179049
O-Subject: [Pegas 1.0 BZ1448160 PATCH V2 23/88] kvm: arm/arm64: Rework gpa callback handlers
Bugzilla: 1448160
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Suzuki K Poulose <suzuki.poulose@arm.com>

In order to perform an operation on a gpa range, we currently iterate
over each page in a user memory slot for the given range. This is
inefficient while dealing with a big range (e.g, a VMA), especially
while unmaping a range. At present, with stage2 unmap on a range with
a hugepage backed region, we clear the PMD when we unmap the first
page in the loop. The remaining iterations simply traverse the page table
down to the PMD level only to see that nothing is in there.

This patch reworks the code to invoke the callback handlers on the
biggest range possible within the memory slot to to reduce the number of
times the handler is called.

Cc: Marc Zyngier <marc.zyngier@arm.com>
Reviewed-by: Christoffer Dall <cdall@linaro.org>
Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Christoffer Dall <cdall@linaro.org>
(cherry picked from commit 056aad67f836e20568be9b7ca07703a8dea26e0c)
Signed-off-by: Wei Huang <wei@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>