[virt] kvm: arm/arm64: vgic: Get rid of unnecessary process_maintenance operation
Message-id: <1495824903-9192-29-git-send-email-wei@redhat.com>
Patchwork-id: 179053
O-Subject: [Pegas 1.0 BZ1448160 PATCH V2 28/88] KVM: arm/arm64: vgic: Get rid of unnecessary process_maintenance operation
Bugzilla: 1448160
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Christoffer Dall <christoffer.dall@linaro.org>

Since we always read back the LRs that we wrote to the guest and the
MISR and EISR registers simply provide a summary of the configuration of
the bits in the LRs, there is really no need to read back those status
registers and process them.  We might as well just signal the
notifyfd when folding the LR state and save some cycles in the process.
We now clear the underflow bit in the fold_lr_state functions as we only
need to clear this bit if we had used all the LRs, so this is as good a
place as any to do that work.

Conflicts:
  * virt/kvm/arm/vgic/vgic-v2.c
  Conflicts caused by 5b0d2cc2, please see upstream discussion at
  https://lkml.org/lkml/2017/4/10/1

Reviewed-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>
(cherry picked from commit af0614991ab64a55f86cda257cedff1be4e435fa)
Signed-off-by: Wei Huang <wei@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>