[arm] soft-reboot into same mode that we entered the kernel
Message-id: <1495824903-9192-56-git-send-email-wei@redhat.com>
Patchwork-id: 179083
O-Subject: [Pegas 1.0 BZ1448160 PATCH V2 55/88] ARM: soft-reboot into same mode that we entered the kernel
Bugzilla: 1448160
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Russell King <rmk+kernel@armlinux.org.uk>

When we soft-reboot (eg, kexec) from one kernel into the next, we need
to ensure that we enter the new kernel in the same processor mode as
when we were entered, so that (eg) the new kernel can install its own
hypervisor - the old kernel's hypervisor will have been overwritten.

In order to do this, we need to pass a flag to cpu_reset() so it knows
what to do, and we need to modify the kernel's own hypervisor stub to
allow it to handle a soft-reboot.

As we are always guaranteed to install our own hypervisor if we're
entered in HYP32 mode, and KVM will have moved itself out of the way
on kexec/normal reboot, we can assume that our hypervisor is in place
when we want to kexec, so changing our hypervisor API should not be a
problem.

Tested-by: Keerthy <j-keerthy@ti.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <cdall@linaro.org>
(cherry picked from commit 9da5ac236de6ab2189c999eb9ddddeef1431ab68)
Signed-off-by: Wei Huang <wei@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>