[virt] kvm: arm/arm64: Cleanup the arch timer code's irqchip checking
Message-id: <1495824903-9192-78-git-send-email-wei@redhat.com>
Patchwork-id: 179101
O-Subject: [Pegas 1.0 BZ1448160 PATCH V2 77/88] KVM: arm/arm64: Cleanup the arch timer code's irqchip checking
Bugzilla: 1448160
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Christoffer Dall <christoffer.dall@linaro.org>

Currently we check if we have an in-kernel irqchip and if the vgic was
properly implemented several places in the arch timer code.  But, we
already predicate our enablement of the arm timers on having a valid
and initialized gic, so we can simply check if the timers are enabled or
not.

This also gets rid of the ugly "error that's not an error but used to
signal that the timer shouldn't poke the gic" construct we have.

Reviewed-by: Alexander Graf <agraf@suse.de>
Reviewed-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>
(cherry picked from commit b22e7df2d85fcbe8b36bab909b98c3d0239e69e6)
Signed-off-by: Wei Huang <wei@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>