[virt] kvm: arm/arm64: vgic-v3: De-optimize VMCR save/restore when emulating a GICv2
Message-id: <1495824903-9192-84-git-send-email-wei@redhat.com>
Patchwork-id: 179105
O-Subject: [Pegas 1.0 BZ1448160 PATCH V2 83/88] KVM: arm/arm64: vgic-v3: De-optimize VMCR save/restore when emulating a GICv2
Bugzilla: 1448160
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Marc Zyngier <marc.zyngier@arm.com>

When emulating a GICv2-on-GICv3, special care must be taken to only
save/restore VMCR_EL2 when ICC_SRE_EL1.SRE is cleared. Otherwise,
all Group-0 interrupts end-up being delivered as FIQ, which is
probably not what the guest expects, as demonstrated here with
an unhappy EFI:

	FIQ Exception at 0x000000013BD21CC4

This means that we cannot perform the load/put trick when dealing
with VMCR_EL2 (because the host has SRE set), and we have to deal
with it in the world-switch.

Fortunately, this is not the most common case (modern guests should
be able to deal with GICv3 directly), and the performance is not worse
than what it was before the VMCR optimization.

Reviewed-by: Christoffer Dall <cdall@linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <cdall@linaro.org>
(cherry picked from commit ff567614d58551b650a2375b50be368fbfed5cd5)
Signed-off-by: Wei Huang <wei@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>