[powerpc] 64s: Idle POWER8 avoid full state loss recovery where possible
Message-id: <20170607190812.36443.11864.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 179466
O-Subject: [PATCH PEGAS1.0 BZ1368734 15/15] powerpc/64s: Idle POWER8 avoid full state loss recovery where possible
Bugzilla: 1368734
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1368734

Description:
------------
If not all threads were in winkle, full state loss recovery is not
necessary and can be avoided. A previous patch removed this optimisation
due to some complexity with the implementation. Re-implement it by
counting the number of threads in winkle with the per-core idle state.
Only restore full state loss if all threads were in winkle.

This has a small window of false positives right before threads execute
winkle and just after they wake up, when the winkle count does not
reflect the true number of threads in winkle. This is not a significant
problem in comparison with even the minimum winkle duration. For
correctness, a false positive is not a problem (only false negatives
would be).

Reviewed-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 0d7720a24234cefa8b318b1a3cf0b302a51faec9

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=0d7720a24234cefa8b318b1a3cf0b302a51faec9

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>