[virt] KVM: arm64: vgic-its: Fix pending table sync
Message-id: <1496996865-29062-24-git-send-email-eric.auger@redhat.com>
Patchwork-id: 179570
O-Subject: [Pegas 1.0 BZ1411488 PATCH 23/37] KVM: arm64: vgic-its: Fix pending table sync
Bugzilla: 1411488
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

In its_sync_lpi_pending_table() we currently ignore the
target_vcpu of the LPIs. We sync the pending bit found in
the vcpu pending table even if the LPI is not targeting it.

Also in vgic_its_cmd_handle_invall() we are supposed to
read the config table data for the LPIs associated to the
collection ID. At the moment we refresh all LPI config
information.

This patch passes a vpcu to vgic_copy_lpi_list() so that
this latter returns a snapshot of the LPIs targeting this
CPU and only those.

Signed-off-by: Eric Auger <eric.auger@redhat.com>
Reviewed-by: Christoffer Dall <cdall@linaro.org>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
(cherry picked from commit ccb1d791ab9e9d45be15a779e6a42fc661309af1)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>